backend/src/main.py
import sys
import logging
import threading
import time
import multiprocessing
from fastapi import FastAPI
from fastapi.staticfiles import StaticFiles
from fastapi.middleware.cors import CORSMiddleware
import glob

from core.slskd_manager import SlskcManager
from core.library_service import LibraryService
from core.metadata_service import MetadataService
from core.romanization_service import RomanizationService
from core.song_processor import SongProcessor
from core.playlist_service import PlaylistService
from api import search_routes, download_routes, library_routes, system_routes, playlist_routes
from api.search_routes import router as search_router
from core.config_utils import get_config_path, get_documents_folder

import os
import json
from configparser import ConfigParser

from dotenv import load_dotenv
load_dotenv()

def create_default_config_if_not_exists():
    """Create a default config file if one doesn't exist."""
    config_path = get_config_path()
    if not os.path.exists(config_path):
        config = ConfigParser()
        documents_folder = get_documents_folder()
        default_data_path = os.path.join(documents_folder, "Sonosano_Songs")
        os.makedirs(default_data_path, exist_ok=True)
        
        config['Paths'] = {'dataPath': default_data_path}
        with open(config_path, 'w') as f:
            config.write(f)

def load_data_path():
    """Load the data path from the config file."""
    create_default_config_if_not_exists()
    config_path = get_config_path()
    config = ConfigParser()
    config.read(config_path)
    if 'Paths' in config and 'dataPath' in config['Paths']:
        return config['Paths']['dataPath']
    return None

if getattr(sys, 'frozen', False):
    multiprocessing.freeze_support()
    application_path = os.path.dirname(sys.executable)
else:
    application_path = os.path.dirname(os.path.abspath(__file__))

data_path = load_data_path()
if data_path is None:
    documents_folder = get_documents_folder()
    data_path = os.path.join(documents_folder, "Sonosano_Songs")
    os.makedirs(data_path, exist_ok=True)

logging.info(f"Using data folder at: {data_path}")

app = FastAPI(title="Sonosano API", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

metadata_service = MetadataService(data_path)
library_service = LibraryService(metadata_service, data_path)
romanization_service = RomanizationService()
song_processor = SongProcessor(library_service, metadata_service, romanization_service, data_path)
slskd_manager = SlskcManager(library_service, data_path)
playlist_service = PlaylistService(data_path)

search_routes.slskd_manager = slskd_manager
download_routes.slskd_manager = slskd_manager
library_routes.library_service = library_service
playlist_routes.library_service = library_service
playlist_routes.playlist_service = playlist_service
library_routes.song_processor = song_processor
system_routes.slskd_manager = slskd_manager
system_routes.romanization_service = romanization_service
system_routes.data_path = data_path

app.include_router(search_router, prefix="", tags=["search"])
app.include_router(download_routes.router, prefix="", tags=["download"])
app.include_router(library_routes.router, prefix="", tags=["library"])
app.include_router(system_routes.router, prefix="", tags=["system"])
app.include_router(playlist_routes.router, prefix="", tags=["playlists"])

covers_path = os.path.join(data_path, "covers")
os.makedirs(covers_path, exist_ok=True)
app.mount("/covers", StaticFiles(directory=covers_path), name="covers")

temp_path = os.path.join(data_path, "temp")
os.makedirs(temp_path, exist_ok=True)
app.mount("/temp", StaticFiles(directory=temp_path), name="temp")

from watchdog.observers import Observer
from core.file_watcher import MusicFileHandler

def long_running_startup_tasks():
    slskd_manager.initialize_slskd()
    
    def initial_sync():
        logging.info("Waiting for Soulseek login before initial sync...")
        if not slskd_manager.wait_for_login(timeout=60):
            logging.warning("Soulseek login timed out. Initial sync may be incomplete.")
            return
            
        logging.info("=== Starting initial library sync and processing ===")
        all_songs_in_db = {s['path'] for s in library_service.get_all_songs()}
        download_dir = os.path.join(data_path, "downloads")
        logging.info(f"Scanning music directory for sync: {download_dir}")
        
        fs_files = set()
        for ext in {'.mp3', '.wav', '.flac', '.m4a', '.aac', '.ogg', '.wma', '.opus'}:
            # Use relative paths for comparison
            for file_path in glob.glob(os.path.join(download_dir, f"**/*{ext}"), recursive=True):
                fs_files.add(os.path.relpath(file_path, download_dir))

        new_files = fs_files - all_songs_in_db
        deleted_files = all_songs_in_db - fs_files
        logging.info(f"Sync found {len(new_files)} new files and {len(deleted_files)} deleted files.")

        for file_path in deleted_files:
            logging.info(f"Sync: Removing deleted song from DB: {file_path}")
            library_service.remove_song(file_path)

        logging.info(f"Processing {len(new_files)} new files for metadata, covers, and lyrics.")
        for file_path in new_files:
            if file_path not in song_processor._currently_processing:
                # Reconstruct the absolute path for the song processor
                abs_path = os.path.join(download_dir, file_path)
                song_processor.process_new_song(abs_path)
        logging.info("=== Initial library sync and processing finished. ===")

    sync_thread = threading.Thread(target=initial_sync, daemon=True)
    sync_thread.start()

    def start_watcher():
        download_dir = os.path.join(data_path, "downloads")
        os.makedirs(download_dir, exist_ok=True)
        event_handler = MusicFileHandler(library_service, song_processor, download_dir)
        observer = Observer()
        observer.schedule(event_handler, download_dir, recursive=True)
        observer.start()
        try:
            while True: time.sleep(1)
        except KeyboardInterrupt:
            observer.stop()
        observer.join()

    watcher_thread = threading.Thread(target=start_watcher, daemon=True)
    watcher_thread.start()

@app.on_event("startup")
async def startup_event():
    startup_thread = threading.Thread(target=long_running_startup_tasks, daemon=True)
    startup_thread.start()

def main():
    import uvicorn
    import logging

    logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
    uvicorn.run(app, host="0.0.0.0", port=8000)

if __name__ == "__main__":
    main()